requirements.txt
fastapi
uvicorn[standard]
transformers
torch
torchvision
Pillow
numpy
httpx
python-multipart
python-dotenv
pydantic